feat(auth): move db cleanup to auth
