+++ packages/auth/src/config/app.ts
@@ -43,5 +43,7 @@ export const Config = {
   authServerDomain: envString('AUTH_SERVER_DOMAIN', 'http://localhost:3006'), // TODO: replace this with whatever frontend port ends up being
   waitTimeSeconds: envInt('WAIT_SECONDS', 5),
   cookieKey: envString('COOKIE_KEY', crypto.randomBytes(32).toString('hex')),
-  accessTokenExpirySeconds: envInt('ACCESS_TOKEN_EXPIRY_SECONDS', 10 * 60) // Default 10 minutes
+  accessTokenExpirySeconds: envInt('ACCESS_TOKEN_EXPIRY_SECONDS', 10 * 60), // Default 10 minutes
+  databaseCleanupWorkers: envInt('DATABASE_CLEANUP_WORKERS', 1),
+  accessTokenCleanupMinAge: 1000 * 60 * 60 * 24 * 30 // 30 days
 }
